- frame_cache     Some changes to the interface, also added missing allocator which was not stored. This may be subject to change in the future.
- pdb
    Fixed a bug where backbone residues not properly computed due to chains not being fully formed.

- script
    Now ir and eval exposes a fingerprint which one can compare against to see if the representation is up to date.
    Split eval procedure into 2 procedures: alloc and compute.
    Fixed array initialization so now it properly concatenates two bitfield arrays. Further testing is required.
    Added interrupt signaling functionality to abort a compute operation if running async.
    Fixed operation precedence balancing bug.
    Fixed bug in PRINT macro which caused the signed size arg to snprintf to be a huge unsigned value singe snprintf accepts size_t
    Fixed expand_node_token_range_with_children which were not called at the right location and thus not giving the intended effects.
    Identifiers are still a bit of a mess and needs a good cleanup, but it seems to work for now.

-util
    Changed to using internal vec3_t types for distance operations in computing secondary structures.

-os
    Added files for dealing with operating system or platform specific functionality... This may be subject to change, timing should perhaps not be a part of this, but is also very small as standalone functionality.

- math
    Moved mat4_inverse from being inlined to a proper function. This was to try and reduce the binary size of mold which seems to have blown up... requires further investigation. My hunch is that something gets inlined alot!
    Binary is close to 20MB, but compresses very well, so there is alot of redundant information (possibly due to inlining which would support the case of too much inlining).

- misc
    Minor cleanup

+++ src/core/md_stack_allocator.c
@@ -17,7 +17,7 @@ void* stack_realloc(md_stack_allocator_t* stack_alloc, void* ptr, uint64_t old_s
         if (ptr == stack_alloc->buf + stack_alloc->prev) {
             const int64_t diff = (int64_t)new_size - (int64_t)old_size;
             const int64_t new_curr = stack_alloc->curr + diff;
-            ASSERT(0 <= new_curr && new_curr < stack_alloc->capacity);
+            ASSERT(0 <= new_curr && new_curr < (int64_t)stack_alloc->capacity);
             stack_alloc->curr = new_curr;
             return ptr;
         }
